+++ lepiter-core/src/lib.rs
@@ -1256,6 +1256,8 @@ pub fn is_code_snippet(typ: &str) -> bool {
             | "exampleSnippet"
             | "changesSnippet"
             | "robocoderMetamodelSnippet"
+            | "rustSnippet"
+            | "tomlSnippet"
     )
 }